+++ flumine/order/trade.py
@@ -158,12 +158,15 @@ class Trade:
     @property
     def info(self) -> dict:
         return {
-            "id": self.id,
-            "strategy": self.strategy,
-            "status": self.status,
+            "id": str(self.id),
+            "strategy": str(self.strategy),
+            "place_reset_seconds": self.place_reset_seconds,
+            "reset_seconds": self.reset_seconds,
             "orders": [o.id for o in self.orders],
             "notes": self.notes_str,
             "market_notes": self.market_notes,
+            "status": self.status.value if self.status else None,
+            "status_log": ", ".join([s.value for s in self.status_log]),
         }
 
     def __enter__(self):